--- instrumentation/khttp-0.1/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/khttp/KHttpAdvice.java
@@ -1,50 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.khttp;
-
-import static io.opentelemetry.javaagent.instrumentation.api.Java8BytecodeBridge.currentContext;
-import static io.opentelemetry.javaagent.instrumentation.khttp.KHttpHeadersInjectAdapter.asWritable;
-import static io.opentelemetry.javaagent.instrumentation.khttp.KHttpTracer.tracer;
-
-import io.opentelemetry.context.Context;
-import io.opentelemetry.context.Scope;
-import java.util.Map;
-import khttp.responses.Response;
-import net.bytebuddy.asm.Advice;
-
-public class KHttpAdvice {
-
-  @Advice.OnMethodEnter(suppress = Throwable.class)
-  public static void methodEnter(
-      @Advice.Argument(value = 0) String method,
-      @Advice.Argument(value = 1) String uri,
-      @Advice.Argument(value = 2, readOnly = false) Map<String, String> headers,
-      @Advice.Local("otelContext") Context context,
-      @Advice.Local("otelScope") Scope scope) {
-    Context parentContext = currentContext();
-    if (!tracer().shouldStartSpan(parentContext)) {
-      return;
-    }
-
-    headers = asWritable(headers);
-    context = tracer().startSpan(parentContext, new RequestWrapper(method, uri, headers), headers);
-    scope = context.makeCurrent();
-  }
-
-  @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-  public static void methodExit(
-      @Advice.Return Response response,
-      @Advice.Thrown Throwable throwable,
-      @Advice.Local("otelContext") Context context,
-      @Advice.Local("otelScope") Scope scope) {
-    if (scope == null) {
-      return;
-    }
-
-    scope.close();
-    tracer().endMaybeExceptionally(context, response, throwable);
-  }
-}
