+++ instrumentation/khttp-0.1/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/khttp/KHttpInstrumentationModule.java
@@ -5,25 +5,12 @@
 
 package io.opentelemetry.javaagent.instrumentation.khttp;
 
-import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.extendsClass;
-import static io.opentelemetry.javaagent.extension.matcher.ClassLoaderMatcher.hasClassesNamed;
 import static java.util.Collections.singletonList;
-import static java.util.Collections.singletonMap;
-import static net.bytebuddy.matcher.ElementMatchers.isAbstract;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.not;
-import static net.bytebuddy.matcher.ElementMatchers.returns;
-import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 
 import com.google.auto.service.AutoService;
 import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import java.util.List;
-import java.util.Map;
-import net.bytebuddy.description.method.MethodDescription;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
 
 @AutoService(InstrumentationModule.class)
 public class KHttpInstrumentationModule extends InstrumentationModule {
@@ -37,28 +24,4 @@ public class KHttpInstrumentationModule extends InstrumentationModule {
     return singletonList(new KHttpInstrumentation());
   }
 
-  public static class KHttpInstrumentation implements TypeInstrumentation {
-    @Override
-    public ElementMatcher<ClassLoader> classLoaderOptimization() {
-      return hasClassesNamed("khttp.KHttp");
-    }
-
-    @Override
-    public ElementMatcher<TypeDescription> typeMatcher() {
-      return extendsClass(named("khttp.KHttp"));
-    }
-
-    @Override
-    public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-      return singletonMap(
-          isMethod()
-              .and(not(isAbstract()))
-              .and(named("request"))
-              .and(takesArgument(0, String.class))
-              .and(takesArgument(1, String.class))
-              .and(takesArgument(2, Map.class))
-              .and(returns(named("khttp.responses.Response"))),
-          KHttpAdvice.class.getName());
-    }
-  }
 }
